make the requested changes
