+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/playWorkout/component/ColumnWithBackgroundImage.kt
@@ -19,7 +19,7 @@ import org.jetbrains.compose.resources.painterResource
 
 
 @Composable
-fun ColumnWithBackground(
+fun ColumnWithBackgroundImage(
     model: String,
     contentDescription: String?,
     modifier: Modifier = Modifier,
