+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/register/content/RegisterScreenContentSelectHeightAndWeight.kt
@@ -136,7 +136,7 @@ fun RegisterScreenContentSelectHeightAndWeight(
     }
 }
 
-@Preview()
+@Preview
 @Composable
 private fun RegisterScreenContentSelectHeightAndWeightPreview() {
     AppTheme(
