+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/viewmodel/playWorkout/PlayWorkoutViewModel.kt
@@ -109,7 +109,8 @@ class PlayWorkoutViewModel(
         }
         updateState {
             it.copy(
-                stage = Stage.REST
+                stage = Stage.REST,
+                currentStep = nextStep
             )
         }
     }
